requirements.txt
python-binance>=1.0.29
numpy>=2.3.1
requests>=2.32.4
colorama>=0.4.6
websocket-client>=1.7.0
